Add RecentLocation menu, Remove embedded libraries
A RecentLocation menu was implemented which stores the last 10 locations
you teleported to. Futhermore the embedded libraries were removed,
because the copy_dylibs.py script already does this job. I added a
Homebrew scheme which does not execute the copy_dylibs.py script. I plan
to allow installation via Homebrew cask to link against the correct
library versions (If this is possible).