Reorder recent locations
When you teleport to a location which is already in the list of recent
locations, the item will be moved to the beginning of the list.
Futhermore the OpenSSL version was upgraded to 1.1, so that only one
version of the dynamic libraries needs to be bundled. I modified the
schemes to either bundle the dynamic libraries or depened on them. See
the Readme.md for more details.