Add ProgressAlert runSheetModal
The runSheetModal function makes presenting the ProgressAlert block
until the alert is dismissed. This has some downsides, e.g the MapView
will not update because it uses the same event loop as the modal. The
advantage is a cleaner code inside the MapViewController and
WindowController. I'm still undecided on this change.

Note: The Downloader class must not include DispatchQueue.main.async
calls, otherwise this change will not work.