Add support for device pairing
libimobiledevice automatically detects devices, even if they are not
paired. MacOS will automatically prompt you to trust this computer to
initiate the pairing. That means all there is left to do, is to reupload
the DeveloperDiskImage after the pairing was successfull, but only if
the previously unpaired device is currently selected. This can easily be
done by reapplying the current selection, since there can not be any
spoofed location state which we would need to preserve.