Add custom speed slider
This commit adds a speed slider to the toolbar items. You have to
manually customize the toolbar and add the item. This item does not
replace the movement type selection. While the movement type influences
the default speed value, it does also influence the routing algorithm.
E.g. Some routes are only available via car and not for walking or cycling.