Decrease the update interval
Decreasing the update interval from 1.0 seconds to 0.05 seconds. This
will increase the smoothness on fast connections (iOS Simulator or USB
connection). On slow connections this will have no effect, since the
time delta will be used, which is still around ~1.0 - 1.5 seconds. This
might cause glitches in the map overlay, if the update interval is to
small. We will see if any issues are reported.