Add a custom NavigationRenderer and Overlay
For future needs we need a MKOverlayRenderer that also shows the
previous, already traveled route in addition to the upcoming route. To
accomplish this goal a custom NavigationRenderer with a
corresponding NavigationOverlay was implemented. This overlay allows
setting an active and inactive route, that can optinally be decorated
with a border. While this implementation is already heavily optimized
(multithreaded, optimized to only draw the current tile skipping
unnecessary waypoints...), it is still a little bit slower than adding
and removing an MKPolylineRenderer. Nevertheless this implementation is
more elegant and flexible, since it does not require removing and
readding the overlay over and over again. To minimize the load on the
CPU it is a good idea to increase the `kAutoMoveDuration` of
`LocationSpoofer`, since each move  will lead to a redraw of the current
route.