v0.1.9
This update moves the LocationSpoofer backend to its own package. By
moving the backend out of the project, arm64 is not natively supported.
This update adds the `move when standing still` functionality, as well
as an option to autoreverse a navigation. Additionally the UI does no
longer block, when the DeveloperDiskImage is uploaded. A couple of
smaller bug fixes are included as well. A new custom renderer is used
instead of apples build in polyline renderer.