Add Speed menu item + Bug fixes
This fixes some long standing bugs e.g. the wrong menubar items are activated on app launch and if a navigation is reset. Coordinate selection crashes the app if the textfield is cleared. CLLocationSpeed is used in more places to indicate the unit of the speed in m/s. An CLLocationSpeed extension helps with converting between km/h and m/s.

Merge

Add missing files